discord_bot/requirements.txt
discord.py==2.4.0
playwright==1.52.0
aiosqlite==0.20.0
cryptography==43.0.1
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
